requirements.txt
Flask==3.0.3
Flask-PyMongo==2.3.0
dnspython==2.7.0
gunicorn==23.0.0
python-dotenv
